+++ tutoraspects/__about__.py
@@ -2,4 +2,4 @@
 Expose some package metadata.
 """
 
-__version__ = "0.22.0"
+__version__ = "0.24.0"